prj structure upd
